requirements.txt
# 核心依赖
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0

# 扩散模型相关
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.4.0

# 图像处理和评估
scikit-image>=0.19.0
lpips>=0.1.4

# 数据处理和工具
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0

# 可选依赖（用于更好的性能）
xformers>=0.0.20  # 用于更快的注意力计算
safetensors>=0.3.0  # 用于更安全的模型保存

# 开发依赖（可选）
jupyter>=1.0.0
ipython>=8.0.0 